Forsaken Ruins JSON fix
Fixes the itemremove, where the armor items weren't set correctly, as well as removes the items to modify the world (shovel, axe, planks) since you can't build on the map at all. Why give the players the loadout if they can't use it? Also, the golden apple as well as the steaks were both set on slot 8 before...

Tested.