Add Curabitur
I felt like this map would be a nice Avicus map to have on Warzone, I hope the players will like it.

The region files are pruned with mcaselector.

I edited the original kit slightly and removed 2 of the 3 golden apples, changed the 32 logs to 2x 64 oak planks and increased the steaks from 8 to 64, to make it fit Warzone more.

Tested.