Fix Warlock (TM)
Changes the spawn protection on both Warlock DTW and Warlock TM slightly to allow placing blocks right at the spawn. This solves the problem of otherwise having a one-wide gap between the spawn and the monument where you can't place blocks (I'm sure some are familiar with this)

On Warlock TM, the random stone blocks in the air have been removed. The itemremove has been adjusted to what it should have done already: The only pickaxe-minable blocks on the map should be the bricks in the middle of the map. This encourages the teams to fight for the bricks, since bricks can be useful for the defense. No other blocks (sandstone, cobblestone, quartz etc.) can be used to defend this way on this tournament specific map.

It has also been loaded into 1.16.5 and pruned afterwards.