Add Rampant
Tested. The flags on this map don't have patterns, therefore the "patterns" list in the flag object is empty. There is a fix I made for TGM which makes patterns optional, however, this map will probably be merged before the fix is active, therefore I left it with an empty list instead of removing it.

I will remove the redundant patterns list once the TGM fix is merged and active on Warzone.