Update Electri Mini void filter
The JSON used for this map is nearly the same as Electri City. In it, the void filter is intentionally set up to allow block placement in a small portion of two of the wool rooms. If the original JSON were to deny block placement in this small area, then it would interfere with bridging from the nearby island to the wool room.

Since Electri Mini lacks the problematic wool rooms, the void filter has been updated to deny block placement entirely on the sides of both wool rooms.