Add Dead Waters (micro map, though)
I had it almost done anyways once I confirmed it was intended to be for micro matches, but I decided just to finish it just for the sake of adding another map staff can use to switch it up as well as Hacktoberfest 2021.
I tested the map and it seems fine, but one little thing I noticed was that I couldn't for the life of me respawn facing the left exit on the blue team. Might've just been ridiculous odds, but it's worth someone looking into eventually.
Regardless it's not crucial by any means but I thought it was worth noting.