Add Pompeii by kroest & GoldenOre
First map to use the random spawn module https://github.com/Warzone/TGM/pull/760
Spawns are a little sketch, as players are spawn at y=70, resulting in a lot of falling spawns (fall damage is turned off as a result). This is to prevent players from spawning in the ground. If spawns are an issue, I'll manually set a bunch of spawns.