Monument Modes - Change SFX for 1.8 parity
In older versions of the game, the beacon did not have special sound effects. Because of this, players participating on older versions of the game don't receive audible notifications of the changes to the monuments.
This changes the beacon sound effect to the elder guardian curse sound effect (the one you get when it curses you with mining fatigue) which is not only available for 1.8 players, but also just sounds more dramatic :)