Add Creep Slap! by AtlasGames
TGM doesn't register creepers as kills. Therefore the map couldn't work as a TDM. Blitz is used instead.

Kill-rewards don't work. Generators are used instead to supply team's with additional creeper spawn eggs. The rate and amount that are given can be changed. An optional "fast mode" could be put in place near the end of the match, generators would be upgraded to pump out more spawn eggs.

This is not the final product of the map, testing needs to be done to fine-tune.